requirements.txt
torch
torchvision
tensorboard
numpy<2
pillow
